Rename #20 universityselection
