+++ domain/universityselection/src/main/java/com/hankki/domain/universityselection/UniversitySelectionModel.kt
@@ -1,4 +1,4 @@
-package com.hankki.domain.navigationselection
+package com.hankki.domain.universityselection
 
 data class UniversitySelectionModel(
     val id: Int,
